requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.3
pillow~=11.2.1
matplotlib~=3.10.1
scikit-image~=0.25.2
requests~=2.32.3
tqdm~=4.67.1
joblib~=1.5.0
scikit-learn~=1.6.1